refactor : 비즈니스 로직을 갖는 객체를 domain 패키지 내부로 이동
